+++ dotnet/GraphQL/Query.cs
@@ -152,6 +152,56 @@ namespace ReviewsRatings.GraphQL
                 }
             );
 
+            FieldAsync<SearchResponseType>(
+                "reviewByreviewDateTime",
+                arguments: new QueryArguments(
+                    new QueryArgument<NonNullGraphType<StringGraphType>> { Name = "reviewDateTime", Description = "Review DateTime" },
+                    new QueryArgument<StringGraphType> { Name = "orderBy", Description = "Order by" },
+                    new QueryArgument<StringGraphType> { Name = "status", Description = "Status" }
+                ),
+                resolve: async context =>
+                {
+                    string reviewDateTime = context.GetArgument<string>("reviewDateTime");
+                    string orderBy = context.GetArgument<string>("orderBy");
+                    string status = context.GetArgument<string>("status");
+
+                    var searchResult = await productReviewService.GetReviewsByreviewDateTime(reviewDateTime);
+                    SearchResponse searchResponse = new SearchResponse
+                    {
+                        Data = new DataElement { data = searchResult.Reviews },
+                        Range = searchResult.Range
+                    };
+
+                    return searchResponse;
+                }
+            );
+
+              FieldAsync<SearchResponseType>(
+                "reviewByDateRange",
+                arguments: new QueryArguments(
+                    new QueryArgument<StringGraphType> { Name = "fromDate", Description = "From Date" },
+                    new QueryArgument<StringGraphType> { Name = "toDate", Description = "To Date" },
+                    new QueryArgument<StringGraphType> { Name = "orderBy", Description = "Order by" },
+                    new QueryArgument<StringGraphType> { Name = "status", Description = "Status" }
+                ),
+                resolve: async context =>
+                {
+                    string fromDate = context.GetArgument<string>("fromDate");
+                    string toDate = context.GetArgument<string>("toDate");
+                    string orderBy = context.GetArgument<string>("orderBy");
+                    string status = context.GetArgument<string>("status");
+
+                    var searchResult = await productReviewService.GetReviewsByDateRange(fromDate, toDate);
+                    SearchResponse searchResponse = new SearchResponse
+                    {
+                        Data = new DataElement { data = searchResult.Reviews },
+                        Range = searchResult.Range
+                    };
+
+                    return searchResponse;
+                }
+            );
+
             FieldAsync<BooleanGraphType>(
                 "hasShopperReviewed",
                 arguments: new QueryArguments(
@@ -174,6 +224,7 @@ namespace ReviewsRatings.GraphQL
                         async c => await productReviewService.GetAppSettings());
                 }
             );
+
         }
     }
 }