+++ react/DownloadReviewsTable.tsx
@@ -1,0 +1,3 @@
+import DownloadReviewsTable from './admin/DownloadReviewsTable'
+
+export default DownloadReviewsTable